Access control new structure! (#61)
* Add password class and test case for password

* update password class and test to fix error in automated tests

* update test for coverage

* update test for coverage

* implement password for add command.

* update new structure for access control using password

* Enhance code and fix multiple bug

* update Test cases to fit new command structure

* fix coverage and add test cases

* UPDATE accress control structure.

* update minor bug fix

* test case to fit the new access control structure

* Merge branch 'master' of https://github.com/CS2103JAN2018-W14-B2/main into AccessControl

# Conflicts:
#	src/main/java/seedu/ptman/logic/commands/AddCommand.java
#	src/main/java/seedu/ptman/logic/commands/DeleteCommand.java
#	src/main/java/seedu/ptman/logic/commands/EditCommand.java
#	src/main/java/seedu/ptman/logic/commands/RedoCommand.java
#	src/main/java/seedu/ptman/logic/commands/UndoCommand.java
#	src/main/java/seedu/ptman/logic/parser/AddCommandParser.java
#	src/main/java/seedu/ptman/logic/parser/DeleteCommandParser.java
#	src/main/java/seedu/ptman/logic/parser/PartTimeManagerParser.java
#	src/main/java/seedu/ptman/model/Model.java
#	src/main/java/seedu/ptman/model/ModelManager.java
#	src/main/java/seedu/ptman/model/PartTimeManager.java
#	src/main/java/seedu/ptman/model/Password.java
#	src/test/java/seedu/ptman/logic/commands/AddCommandTest.java
#	src/test/java/seedu/ptman/logic/parser/AddCommandParserTest.java

* update clear command to execute when provided password

* update test case for clear,redo,undo command parser

* update test

+++ src/main/java/seedu/ptman/logic/commands/DeleteCommand.java
@@ -1,6 +1,7 @@
 package seedu.ptman.logic.commands;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.ptman.logic.parser.CliSyntax.PREFIX_PASSWORD;
 
 import java.util.List;
 import java.util.Objects;
@@ -8,8 +9,10 @@ import java.util.Objects;
 import seedu.ptman.commons.core.Messages;
 import seedu.ptman.commons.core.index.Index;
 import seedu.ptman.logic.commands.exceptions.CommandException;
+import seedu.ptman.model.Password;
 import seedu.ptman.model.employee.Employee;
 import seedu.ptman.model.employee.exceptions.EmployeeNotFoundException;
+import seedu.ptman.model.employee.exceptions.InvalidPasswordException;
 
 /**
  * Deletes a employee identified using it's last displayed index from PTMan.
@@ -21,24 +24,31 @@ public class DeleteCommand extends UndoableCommand {
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
             + ": Deletes the employee identified by the index number used in the last employee listing.\n"
-            + "Parameters: INDEX (must be a positive integer)\n"
-            + "Example: " + COMMAND_WORD + " 1";
+            + "Parameters: INDEX (must be a positive integer) pw/AdminPassword \n"
+            + "Example: " + COMMAND_WORD + " 1 " + PREFIX_PASSWORD + "AdminPassword";
 
     public static final String MESSAGE_DELETE_EMPLOYEE_SUCCESS = "Deleted Employee: %1$s";
 
     private final Index targetIndex;
 
     private Employee employeeToDelete;
+    private Password toCheck;
 
-    public DeleteCommand(Index targetIndex) {
+    public DeleteCommand(Index targetIndex, Password password) {
         this.targetIndex = targetIndex;
-        isAdminCommand = true;
+        toCheck = password;
     }
 
 
     @Override
-    public CommandResult executeUndoableCommand() {
+    public CommandResult executeUndoableCommand() throws InvalidPasswordException {
         requireNonNull(employeeToDelete);
+        requireNonNull(toCheck);
+
+        if (!model.isAdminPassword(toCheck)) {
+            throw new InvalidPasswordException();
+        }
+
         try {
             model.deleteEmployee(employeeToDelete);
         } catch (EmployeeNotFoundException pnfe) {
